# api/openapi.yaml
openapi: 3.1.0
info:
  title: Superserve API
  version: 0.1.0
  description: |
    Superserve provides sandbox infrastructure to run AI agents in the cloud. Powered by Firecracker MicroVMs.

    ## Sandbox lifecycle

    ```
    active <--> paused --> deleted
    ```

    A sandbox is `active` when running and `paused` after being paused. Resuming
    a paused sandbox returns it to `active`. Deleting releases all resources.

    | Endpoint | What it does |
    |----------|-------------|
    | `POST /sandboxes` | Create a new sandbox (optionally `from_template`) |
    | `PATCH /sandboxes/:id` | Partially update a running sandbox (e.g. network rules) |
    | `POST /sandboxes/:id/pause` | Snapshot full state, suspend the VM |
    | `POST /sandboxes/:id/resume` | Restore from snapshot, continue where it left off |
    | `DELETE /sandboxes/:id` | Delete sandbox and all resources |

    ## Sandbox environment

    By default sandboxes boot from the curated `superserve/base` template (Ubuntu 24.04,
    1 vCPU, 1 GB RAM, 4 GB disk, with Python 3.12, Node.js 22, npm, git, curl,
    and build-essential pre-installed). Callers can override with any template
    name (e.g. `superserve/python-3.11`, `superserve/node-22`) or a team-owned template UUID
    via the `from_template` field on `POST /sandboxes`.

    ## Files and commands

    `/files`, `/exec`, and `/exec/stream` run against a single sandbox and use
    its `X-Access-Token` (returned by create, resume, and activate), not the
...

...
servers:
...
paths:
...
  /sandboxes/{sandbox_id}:
    parameters:
      - $ref: "#/components/parameters/SandboxId"

    get:
      operationId: getSandbox
      tags: [Sandboxes]
      summary: Get a sandbox by ID
      security:
        - apiKey: []
      responses:
        "200":
          description: Sandbox details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SandboxResponse"
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      operationId: deleteSandbox
      tags: [Sandboxes]
      summary: Delete a sandbox
      security:
        - apiKey: []
      responses:
        "204":
          description: Sandbox deleted
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      operationId: patchSandbox
      tags: [Sandboxes]
      summary: Partially update a running sandbox
      description: |
        Applies a partial update to a running sandbox. Each top-level field
        in the request body is optional; only fields that are present are
        applied. Omitted top-level fields are left unchanged. Nested objects
        are full replacements when present — to clear a list, send it as an
        empty array.

        At least one top-level field must be present, otherwise the request
        is rejected with `400`. Unknown top-level fields are also rejected
        with `400` so typos surface as errors instead of silent no-ops.

        ## Currently patchable fields

        - `network` — replaces the egress allow/deny rules. The sandbox
          must be in the `active` state; patching a paused sandbox
          returns `409`. Rules take effect immediately and are persisted so
          they survive a future pause/resume cycle.
        - `metadata` — replaces the sandbox's metadata tags. Can be updated
          regardless of sandbox state (active, paused).
        - `auto_delete_seconds` — sets or clears (`null`) the
          garbage-collection window for the paused state. Can be updated
          regardless of sandbox state. When applied to an already-paused
          sandbox, the deletion deadline counts from the moment of this
          request — never retroactively from when the sandbox paused — so
          you always get the full window.
        - `timeout_seconds` — sets or clears (`null`) the auto-pause
          timeout. Can be updated regardless of sandbox state; on a paused
          sandbox it applies to the next active session. The timeout is
          evaluated against the current active session, so lowering it
          below already-elapsed time pauses the sandbox promptly.
        - `preview_access` — moves an older `legacy_public` sandbox to strict
          `public` preview routing. Only explicitly published ports remain
          reachable. The published-port set is unchanged, and the policy
          persists across pause/resume. `legacy_public` cannot be selected.
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SandboxPatch"
      responses:
        "204":
          description: Patch applied
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /sandboxes/{sandbox_id}/preview-ports:
    parameters:
      - $ref: "#/components/parameters/SandboxId"

    get:
      operationId: listSandboxPreviewPorts
      tags: [Sandboxes]
      summary: List published preview ports
      description: |
        Returns the sandbox's preview routing mode and its published ports.
        For a strict `public` sandbox, only the ports listed here are reachable
        at `https://{port}-{sandbox_id}.<sandbox-domain>`; every other numeric
        port returns 404. An older `legacy_public` sandbox may still route all
        listening ports until it is updated to `public`.
      security:
        - apiKey: []
      responses:
        "200":
          description: Preview routing mode and published-port set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewPortList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      operationId: publishSandboxPreviewPort
      tags: [Sandboxes]
      summary: Publish a public preview port
      description: |
        Adds a port to the sandbox's published set. Published ports are public
        in this phase. Publishing is idempotent: publishing an existing port
        again returns the same port and leaves the set unchanged.
      security:
        - apiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishPortRequest"
      responses:
        "200":
          description: Published port
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewPort"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /sandboxes/{sandbox_id}/preview-ports/{port}:
    parameters:
      - $ref: "#/components/parameters/SandboxId"
      - $ref: "#/components/parameters/PreviewPort"

    delete:
      operationId: unpublishSandboxPreviewPort
      tags: [Sandboxes]
      summary: Unpublish a preview port
      description: |
        Removes a port from the published set. On a strict `public` sandbox,
        the port immediately returns 404 at the edge.

        The operation is idempotent and retry-safe. Repeating the request,
        including for a port that is already unpublished, re-synchronizes the
        host's authoritative published-port set and returns `204`. Retry a
        `500` to converge after a transient host update failure. A `404`
        means the sandbox itself does not exist.
      security:
        - apiKey: []
      responses:
        "204":
          description: Port unpublished or already unpublished
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
    accessToken:
      type: apiKey
      in: header
      name: X-Access-Token
      description: >
        Per-sandbox token returned by create, resume, and activate. Grants
        access to a single sandbox's files and commands. Distinct from the
        team `X-API-Key`.
  parameters:
    SandboxId:
      name: sandbox_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/PublicSandboxId"
      description: The unique identifier of the sandbox.
    PreviewPort:
      name: port
      in: path
      required: true
      schema:
        type: integer
        minimum: 1024
        maximum: 65535
      description: Published sandbox port.
...
  schemas:
    PublicSandboxId:
      type: string
      pattern: '^(sb-[a-z0-9]+-)?[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      description: >
        Public sandbox ID: a bare UUID, or the region-tagged form
        `sb-<region>-<uuid>` (e.g. `sb-use-1b4e28ba-…`). Treat as an opaque
        string; the tagged form routes the request to the sandbox's home
        region. Endpoints accept both forms interchangeably.
...
    SandboxListItem:
      description: Sandbox shape returned by list endpoints.
      type: object
      properties:
        id:
          $ref: "#/components/schemas/PublicSandboxId"
        name:
          type: string
        status:
          type: string
          enum: [active, paused, resuming]
          description: >
            Current state of the sandbox. `active` means running; `paused`
            means paused and awaiting resume. `resuming` is a transient
            state observed while the platform restores a paused sandbox
            (e.g. via auto-resume on `/exec`); clients should retry shortly.
        vcpu_count:
          type: integer
        memory_mib:
          type: integer
        snapshot_id:
          type: string
          format: uuid
          description: ID of the latest snapshot (present after a pause).
        created_at:
          type: string
          format: date-time
        timeout_seconds:
          type: integer
          format: int32
          description: >
            Auto-pause timeout in seconds, if configured. Absent when
            auto-pause is disabled.
        auto_delete_seconds:
          type: integer
          format: int32
          description: >
            Garbage-collection window for the paused state, if configured.
            Absent when auto-delete is disabled.
        auto_delete_at:
          type: string
          format: date-time
          description: >
            When the sandbox will be deleted. Present only while the sandbox
            is paused with `auto_delete_seconds` configured. The deadline is
            armed when the sandbox pauses (or when the setting is applied to
            an already-paused sandbox) and cleared on resume.
        network:
          $ref: "#/components/schemas/NetworkConfig"
          description: >
            Current egress allow/deny rules, if any have been configured.
            Absent when the sandbox uses default network settings.
        metadata:
          type: object
          additionalProperties:
            type: string
          description: >
            User-supplied tags attached at creation. Always present —
            sandboxes created without metadata return `{}` rather than
            being absent.
          example:
            env: prod
            owner: agent-7
        preview_access:
          type: string
          enum: [legacy_public, public]
          description: |
            Preview routing mode. `public` routes only explicitly published
            ports. `legacy_public` may be returned for a sandbox created before
            explicit publication and preserves its previous all-port behavior
            until updated to `public`.
    SandboxResponse:
      description: Single-sandbox shape — `SandboxListItem` plus `access_token` and bound secrets.
      allOf:
        - $ref: "#/components/schemas/SandboxListItem"
        - type: object
          properties:
            access_token:
              type: string
              description: |
                Per-sandbox access token for data-plane operations (file
                upload/download, terminal). Pass as the `X-Access-Token`
                header.
            secrets:
              type: array
              description: >
                Credentials bound to this sandbox. Each entry maps an env-var
                name visible to the agent to the secret name it resolves to.
                `revoked=true` when the underlying secret has been soft-deleted
                (the env var still holds the now-useless proxy token).
              items:
                type: object
                required: [env_key, secret_name]
                properties:
                  env_key:
                    type: string
                  secret_name:
                    type: string
                  revoked:
                    type: boolean
                    default: false
...
    NetworkConfig:
      type: object
      description: >
        Egress network rules for a sandbox. `allow_out` accepts CIDRs
        (e.g. `8.8.8.8/32`) and domain names (e.g. `api.openai.com`,
        `*.github.com`). `deny_out` accepts CIDRs only.
        Private ranges (10/8, 172.16/12, 192.168/16, 127/8, 169.254/16)
        are always blocked regardless of rules.
      properties:
        allow_out:
          type: array
          items:
            type: string
          description: CIDRs or domains to allow.
          example: ["api.openai.com", "*.github.com", "8.8.8.8/32"]
        deny_out:
          type: array
          items:
            type: string
          description: CIDRs to deny. Use `0.0.0.0/0` to block all traffic not in `allow_out`.
          example: ["0.0.0.0/0"]
    SandboxPatch:
      type: object
      description: |
        Partial update body for `PATCH /sandboxes/{sandbox_id}`. Each
        top-level field is optional; only fields that are present are
        applied. Omitted fields are left unchanged. Nested objects are
        full replacements when present — to clear a list, send it as an
        empty array.

        At least one top-level field must be set, otherwise the request
        is rejected with `400`. Unknown fields are also rejected with
        `400`.
      properties:
        network:
          allOf:
            - $ref: "#/components/schemas/NetworkConfig"
          description: |
            Replace the sandbox's egress rules. The sandbox must be in
            the `active` state. The provided `allow_out` and `deny_out`
            lists fully replace whatever was previously configured.
        metadata:
          type: object
          additionalProperties:
            type: string
          description: |
            Replace the sandbox's metadata tags. Fully replaces the existing
            metadata — omitted keys are removed. Can be patched regardless of
            sandbox state. Same validation limits as on create (64 keys,
            256-byte keys, 2 KB values, 16 KB total).
        auto_delete_seconds:
          type: integer
          format: int32
          nullable: true
          minimum: 0
          maximum: 2592000
          description: |
            Set or clear the garbage-collection window for the paused state.
            Once the sandbox has been continuously paused for this many
            seconds it is deleted automatically. `0` deletes as soon as the
            sandbox pauses; `null` disables auto-delete. On an already-paused
            sandbox the deadline counts from this request, so the sandbox
            gets the full window. Maximum 2592000 (30 days).
        timeout_seconds:
          type: integer
          format: int32
          nullable: true
          minimum: 1
          maximum: 604800
          description: |
            Set or clear the auto-pause timeout. Same semantics as on
            create; `null` disables auto-pause. Evaluated against the
            current active session, so lowering it below already-elapsed
            time pauses the sandbox promptly. Maximum 604800 (7 days).
        preview_access:
          type: string
          enum: [public]
          description: |
            Move the sandbox to strict public preview routing. Only explicitly
            published ports are reachable after this update. The published-port
            set is unchanged, and the policy persists across pause/resume.
            `legacy_public` cannot be selected through the API.
      example:
        metadata:
          env: prod
          owner: agent-7
    PublishPortRequest:
      type: object
      required: [port]
      additionalProperties: false
      properties:
        port:
          type: integer
          minimum: 1024
          maximum: 65535
          description: Public port to publish.
    PreviewPort:
      type: object
      required: [port]
      properties:
        port:
          type: integer
          minimum: 1024
          maximum: 65535
          description: Published public port.
    PreviewPortList:
      type: object
      required: [preview_access, ports]
      properties:
        preview_access:
          type: string
          enum: [legacy_public, public]
          description: |
            Sandbox preview routing mode. `public` enforces the listed ports;
            `legacy_public` preserves pre-publication routing for older
            sandboxes until they are updated.
        ports:
          type: array
          items:
            $ref: "#/components/schemas/PreviewPort"
...
    Error:
      type: object
      description: |
        Error envelope. `error.code` is a stable, machine-readable identifier
        (e.g. `bad_request`, `not_found`, `conflict`, `rate_limited`,
        `too_many_builds`, `too_many_templates`, `too_many_sandboxes`,
        `image_pull_failed`, `step_failed`, `snapshot_failed`,
        `start_cmd_failed`, `ready_cmd_failed`, `build_failed`).
        `error.message` is human-readable and may change between releases;
        clients should branch on `code`, not `message`.
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing or invalid API key
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Caller is authenticated but not allowed to perform the action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Operation conflicts with the resource's current state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...